Fixed a problem where the game engine's xPrev and yPrev values were being updated too late to be changed for onCollide functions; moved from Object.cpp to Scene.cpp.

+++ source/r1/headers/gameobjects.hpp
@@ -13,8 +13,8 @@ class Player: public Object{
     public:
 		Player(float x, float y, int collisionLayer = 0, unsigned int collisionFlags = 0, bool grav = false);
         void create();
-        void onCollide(Object* other, int myBoxID, int otherBoxID);
-        void process(float delta);
+		void onCollide(Object *other, int myBoxID, int otherBoxID);
+		void process(double delta);
     private:
         std::vector<Sprite*> sprites;
         bool grounded;
